LPS-61231 Fix the fallback and allow the edit page action to handle them if necessary
