+++ modules/apps/wiki/wiki-web/src/main/java/com/liferay/wiki/web/portlet/action/EditPageAttachmentsMVCActionCommand.java
@@ -18,25 +18,19 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.lock.DuplicateLockException;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.portlet.JSONPortletResponseUtil;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
-import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.servlet.ServletResponseConstants;
 import com.liferay.portal.kernel.servlet.SessionErrors;
 import com.liferay.portal.kernel.upload.LiferayFileItemException;
 import com.liferay.portal.kernel.upload.UploadException;
-import com.liferay.portal.kernel.upload.UploadPortletRequest;
 import com.liferay.portal.kernel.upload.UploadRequestSizeException;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.ContentTypes;
-import com.liferay.portal.kernel.util.ObjectValuePair;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PrefsPropsUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
-import com.liferay.portal.kernel.util.StreamUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.TextFormatter;
@@ -48,33 +42,24 @@ import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.PortalUtil;
 import com.liferay.portlet.asset.AssetCategoryException;
 import com.liferay.portlet.asset.AssetTagException;
-import com.liferay.portlet.asset.model.AssetVocabulary;
 import com.liferay.portlet.documentlibrary.DuplicateFileEntryException;
 import com.liferay.portlet.documentlibrary.DuplicateFolderNameException;
 import com.liferay.portlet.documentlibrary.FileExtensionException;
 import com.liferay.portlet.documentlibrary.FileMimeTypeException;
 import com.liferay.portlet.documentlibrary.FileNameException;
 import com.liferay.portlet.documentlibrary.FileSizeException;
-import com.liferay.portlet.documentlibrary.InvalidFileEntryTypeException;
 import com.liferay.portlet.documentlibrary.InvalidFileVersionException;
 import com.liferay.portlet.documentlibrary.NoSuchFileEntryException;
 import com.liferay.portlet.documentlibrary.NoSuchFolderException;
 import com.liferay.portlet.documentlibrary.SourceFileNameException;
 import com.liferay.portlet.documentlibrary.antivirus.AntivirusScannerException;
-import com.liferay.portlet.documentlibrary.model.DLFileEntry;
 import com.liferay.portlet.dynamicdatamapping.StorageFieldRequiredException;
-import com.liferay.portlet.trash.service.TrashEntryService;
 import com.liferay.portlet.trash.util.TrashUtil;
 import com.liferay.taglib.util.RestoreEntryUtil;
+import com.liferay.wiki.WikiAttachmentsHelper;
 import com.liferay.wiki.constants.WikiPortletKeys;
 import com.liferay.wiki.exception.NoSuchNodeException;
 import com.liferay.wiki.exception.NoSuchPageException;
-import com.liferay.wiki.service.WikiPageService;
-
-import java.io.InputStream;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
@@ -89,6 +74,7 @@ import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Jorge Ferrer
+ * @author Roberto Díaz
  */
 @Component(
 	immediate = true,
@@ -102,83 +88,19 @@ import org.osgi.service.component.annotations.Reference;
 )
 public class EditPageAttachmentsMVCActionCommand extends BaseMVCActionCommand {
 
-	protected void addAttachment(ActionRequest actionRequest) throws Exception {
-		UploadPortletRequest uploadPortletRequest =
-			PortalUtil.getUploadPortletRequest(actionRequest);
+	@Reference(unbind = "-")
+	public void set_wikiAttachmentsHelper(
+		WikiAttachmentsHelper wikiAttachmentsHelper) {
 
-		long nodeId = ParamUtil.getLong(actionRequest, "nodeId");
-		String title = ParamUtil.getString(actionRequest, "title");
-
-		int numOfFiles = ParamUtil.getInteger(actionRequest, "numOfFiles");
-
-		List<ObjectValuePair<String, InputStream>> inputStreamOVPs =
-			new ArrayList<>();
-
-		try {
-			if (numOfFiles == 0) {
-				String fileName = uploadPortletRequest.getFileName("file");
-				InputStream inputStream = uploadPortletRequest.getFileAsStream(
-					"file");
-
-				if (inputStream != null) {
-					ObjectValuePair<String, InputStream> inputStreamOVP =
-						new ObjectValuePair<>(fileName, inputStream);
-
-					inputStreamOVPs.add(inputStreamOVP);
-				}
-			}
-			else {
-				for (int i = 1; i <= numOfFiles; i++) {
-					String fileName = uploadPortletRequest.getFileName(
-						"file" + i);
-					InputStream inputStream =
-						uploadPortletRequest.getFileAsStream("file" + i);
-
-					if (inputStream == null) {
-						continue;
-					}
-
-					ObjectValuePair<String, InputStream> inputStreamOVP =
-						new ObjectValuePair<>(fileName, inputStream);
-
-					inputStreamOVPs.add(inputStreamOVP);
-				}
-			}
-
-			_wikiPageService.addPageAttachments(nodeId, title, inputStreamOVPs);
-		}
-		finally {
-			for (ObjectValuePair<String, InputStream> inputStreamOVP :
-					inputStreamOVPs) {
-
-				InputStream inputStream = inputStreamOVP.getValue();
-
-				StreamUtil.cleanUp(inputStream);
-			}
-		}
+		_wikiAttachmentsHelper = wikiAttachmentsHelper;
 	}
 
 	protected void deleteAttachment(
 			ActionRequest actionRequest, boolean moveToTrash)
 		throws Exception {
 
-		long nodeId = ParamUtil.getLong(actionRequest, "nodeId");
-		String title = ParamUtil.getString(actionRequest, "title");
-		String attachment = ParamUtil.getString(actionRequest, "fileName");
-
-		TrashedModel trashedModel = null;
-
-		if (moveToTrash) {
-			FileEntry fileEntry = _wikiPageService.movePageAttachmentToTrash(
-				nodeId, title, attachment);
-
-			if (fileEntry.getModel() instanceof DLFileEntry) {
-				trashedModel = (DLFileEntry)fileEntry.getModel();
-			}
-		}
-		else {
-			_wikiPageService.deletePageAttachment(nodeId, title, attachment);
-		}
+		TrashedModel trashedModel = _wikiAttachmentsHelper.deleteAttachment(
+			actionRequest, moveToTrash);
 
 		if (moveToTrash && (trashedModel != null)) {
 			TrashUtil.addTrashSessionMessages(
@@ -220,7 +142,7 @@ public class EditPageAttachmentsMVCActionCommand extends BaseMVCActionCommand {
 				throw new PortalException(cause);
 			}
 			else if (cmd.equals(Constants.ADD)) {
-				addAttachment(actionRequest);
+				_wikiAttachmentsHelper.addAttachments(actionRequest);
 			}
 			else if (cmd.equals(Constants.CHECK)) {
 				JSONObject jsonObject = RestoreEntryUtil.checkEntry(
@@ -235,16 +157,16 @@ public class EditPageAttachmentsMVCActionCommand extends BaseMVCActionCommand {
 				deleteAttachment(actionRequest, false);
 			}
 			else if (cmd.equals(Constants.EMPTY_TRASH)) {
-				emptyTrash(actionRequest);
+				_wikiAttachmentsHelper.emptyTrash(actionRequest);
 			}
 			else if (cmd.equals(Constants.MOVE_TO_TRASH)) {
 				deleteAttachment(actionRequest, true);
 			}
 			else if (cmd.equals(Constants.RENAME)) {
-				restoreRename(actionRequest);
+				_wikiAttachmentsHelper.restoreRename(actionRequest);
 			}
 			else if (cmd.equals(Constants.RESTORE)) {
-				restoreEntries(actionRequest);
+				_wikiAttachmentsHelper.restoreEntries(actionRequest);
 
 				String redirect = ParamUtil.getString(
 					actionRequest, "redirect");
@@ -254,7 +176,7 @@ public class EditPageAttachmentsMVCActionCommand extends BaseMVCActionCommand {
 				}
 			}
 			else if (cmd.equals(Constants.OVERRIDE)) {
-				restoreOverride(actionRequest);
+				_wikiAttachmentsHelper.restoreOverride(actionRequest);
 			}
 		}
 		catch (NoSuchNodeException | NoSuchPageException |
@@ -268,99 +190,6 @@ public class EditPageAttachmentsMVCActionCommand extends BaseMVCActionCommand {
 			handleUploadException(
 				portletConfig, actionRequest, actionResponse, cmd, e);
 		}
-	}
-
-	protected void emptyTrash(ActionRequest actionRequest) throws Exception {
-		long nodeId = ParamUtil.getLong(actionRequest, "nodeId");
-		String title = ParamUtil.getString(actionRequest, "title");
-
-		_wikiPageService.deleteTrashPageAttachments(nodeId, title);
-	}
-
-	/**
-	 * TODO: Remove. This should extend from EditFileEntryAction once it is
-	 * modularized.
-	 */
-	protected String getAddMultipleFileEntriesErrorMessage(
-			PortletConfig portletConfig, ActionRequest actionRequest,
-			ActionResponse actionResponse, Exception e)
-		throws Exception {
-
-		String errorMessage = null;
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		if (e instanceof AntivirusScannerException) {
-			AntivirusScannerException ase = (AntivirusScannerException)e;
-
-			errorMessage = themeDisplay.translate(ase.getMessageKey());
-		}
-		else if (e instanceof AssetCategoryException) {
-			AssetCategoryException ace = (AssetCategoryException)e;
-
-			AssetVocabulary assetVocabulary = ace.getVocabulary();
-
-			String vocabularyTitle = StringPool.BLANK;
-
-			if (assetVocabulary != null) {
-				vocabularyTitle = assetVocabulary.getTitle(
-					themeDisplay.getLocale());
-			}
-
-			if (ace.getType() == AssetCategoryException.AT_LEAST_ONE_CATEGORY) {
-				errorMessage = themeDisplay.translate(
-					"please-select-at-least-one-category-for-x",
-					vocabularyTitle);
-			}
-			else if (ace.getType() ==
-						AssetCategoryException.TOO_MANY_CATEGORIES) {
-
-				errorMessage = themeDisplay.translate(
-					"you-cannot-select-more-than-one-category-for-x",
-					vocabularyTitle);
-			}
-		}
-		else if (e instanceof DuplicateFileEntryException) {
-			errorMessage = themeDisplay.translate(
-				"the-folder-you-selected-already-has-an-entry-with-this-name." +
-					"-please-select-a-different-folder");
-		}
-		else if (e instanceof FileExtensionException) {
-			errorMessage = themeDisplay.translate(
-				"please-enter-a-file-with-a-valid-extension-x",
-				StringUtil.merge(
-					getAllowedFileExtensions(
-						portletConfig, actionRequest, actionResponse)));
-		}
-		else if (e instanceof FileNameException) {
-			errorMessage = themeDisplay.translate(
-				"please-enter-a-file-with-a-valid-file-name");
-		}
-		else if (e instanceof FileSizeException) {
-			long fileMaxSize = PrefsPropsUtil.getLong(
-				PropsKeys.DL_FILE_MAX_SIZE);
-
-			if (fileMaxSize == 0) {
-				fileMaxSize = PrefsPropsUtil.getLong(
-					PropsKeys.UPLOAD_SERVLET_REQUEST_IMPL_MAX_SIZE);
-			}
-
-			errorMessage = themeDisplay.translate(
-				"please-enter-a-file-with-a-valid-file-size-no-larger-than-x",
-				TextFormatter.formatStorageSize(
-					fileMaxSize, themeDisplay.getLocale()));
-		}
-		else if (e instanceof InvalidFileEntryTypeException) {
-			errorMessage = themeDisplay.translate(
-				"the-document-type-you-selected-is-not-valid-for-this-folder");
-		}
-		else {
-			errorMessage = themeDisplay.translate(
-				"an-unexpected-error-occurred-while-saving-your-document");
-		}
-
-		return errorMessage;
 	}
 
 	/**
@@ -533,70 +362,6 @@ public class EditPageAttachmentsMVCActionCommand extends BaseMVCActionCommand {
 		}
 	}
 
-	protected void restoreEntries(ActionRequest actionRequest)
-		throws Exception {
-
-		long trashEntryId = ParamUtil.getLong(actionRequest, "trashEntryId");
-
-		if (trashEntryId > 0) {
-			_trashEntryService.restoreEntry(trashEntryId);
-
-			return;
-		}
-
-		long[] restoreEntryIds = StringUtil.split(
-			ParamUtil.getString(actionRequest, "restoreTrashEntryIds"), 0L);
-
-		for (long restoreEntryId : restoreEntryIds) {
-			_trashEntryService.restoreEntry(restoreEntryId);
-		}
-	}
-
-	protected void restoreOverride(ActionRequest actionRequest)
-		throws Exception {
-
-		long trashEntryId = ParamUtil.getLong(actionRequest, "trashEntryId");
-
-		long duplicateEntryId = ParamUtil.getLong(
-			actionRequest, "duplicateEntryId");
-
-		_trashEntryService.restoreEntry(trashEntryId, duplicateEntryId, null);
-	}
-
-	protected void restoreRename(ActionRequest actionRequest) throws Exception {
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		long trashEntryId = ParamUtil.getLong(actionRequest, "trashEntryId");
-
-		String newName = ParamUtil.getString(actionRequest, "newName");
-
-		if (Validator.isNull(newName)) {
-			String oldName = ParamUtil.getString(actionRequest, "oldName");
-
-			newName = TrashUtil.getNewName(themeDisplay, null, 0, oldName);
-		}
-
-		_trashEntryService.restoreEntry(trashEntryId, 0, newName);
-	}
-
-	@Reference(unbind = "-")
-	protected void setTrashEntryService(TrashEntryService trashEntryService) {
-		_trashEntryService = trashEntryService;
-	}
-
-	@Reference(unbind = "-")
-	protected void setWikiPageService(WikiPageService wikiPageService) {
-		_wikiPageService = wikiPageService;
-	}
-
-	private static final String _TEMP_FOLDER_NAME =
-		EditPageAttachmentsMVCActionCommand.class.getName();
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		EditPageAttachmentsMVCActionCommand.class);
-
-	private TrashEntryService _trashEntryService;
-	private WikiPageService _wikiPageService;
+	private volatile WikiAttachmentsHelper _wikiAttachmentsHelper;
 
 }
